+++ dart/gui/all.hpp
@@ -4,6 +4,7 @@
 #include <dart/gui/debug.hpp>
 #include <dart/gui/export.hpp>
 #include <dart/gui/geometry.hpp>
+#include <dart/gui/gizmo.hpp>
 #include <dart/gui/interaction.hpp>
 #include <dart/gui/panel.hpp>
 #include <dart/gui/profile.hpp>